+++ Projet 3 Courriers/src/model/content/letter/PromissoryNote.java
@@ -5,8 +5,8 @@ import model.content.Money;
 
 public class PromissoryNote extends Letter<Money> {
 	
-	public PromissoryNote(Inhabitant sender, Inhabitant receiver, double cost,int amount) {
-		super(sender, receiver,new Money(amount), cost);
+	public PromissoryNote(Inhabitant sender, Inhabitant receiver, int cost, int amount) {
+		super(sender, receiver, new Money(amount), cost);
 	}
 	
 	@Override
